Separated Integration test cases
