+++ tests/integration/cases/basic/artifacts/manofest.py
